refactor(trade): split trade.rs into module directory along natural seams
The cell-side trade module grew to 1,895 lines after the trust-boundary
fix-up landed (8 new regression tests). Per CLAUDE.md "File organization":
500-line soft cap, 700-line hard cap. Split along natural seams rather
than arbitrary line counts:

- trade/mod.rs        (49)  module docs + re-exports + MAX_INTERACT_DISTANCE
- trade/handlers.rs   (426) the 4 inbound cell-method handlers + dispatch
- trade/state.rs      (259) session lifecycle (begin/apply/cancel/in-range,
                            cancel_trade_on_disconnect public entry)
- trade/wire.rs       (202) onTradeState/onTradeResults senders + the
                            stub_inv_items_for builder + sentinel rationale
- trade/handoff.rs    (171) request_execute_trade + ExecuteTradeSnapshot
- trade/tests/mod.rs  (47)  shared helpers (make_two_players, build_trade_request_args)
- trade/tests/handlers.rs   (372) request/cancel/update/disconnect happy + reject
- trade/tests/lock_state.rs (176) tradeLockState progression + downgrade + reject
- trade/tests/handoff.rs    (215) both-confirm + distance-recheck regression guard
- trade/tests/wire.rs       (160) stub_inv_items_for sentinel byte fixture

Re-export discipline preserved: `crate::cell::cell_methods::player::trade::dispatch`
and `cancel_trade_on_disconnect` (the two public entry points used by
the dispatch router and base_messages handler) keep their existing paths
via `pub use` in `trade/mod.rs`. No external callsite touched.

Every file is under the 500-line soft cap.
